+++ client/src/layouts/Dashboard/components/Topbar/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Topbar";
