+++ client/src/layouts/Dashboard/components/index.js
@@ -1,0 +1,3 @@
+export { default as Footer } from "./Footer";
+export { default as Sidebar } from "./Sidebar";
+export { default as Topbar } from "./Topbar";
