+++ client/src/layouts/Dashboard/index.js
@@ -1,78 +1,1 @@
-import React, { useState } from "react";
-import clsx from "clsx";
-import { makeStyles, useTheme } from "@material-ui/core/styles";
-import { useMediaQuery } from "@material-ui/core";
-
-import Container from "@material-ui/core/Container";
-import Header from "./Header";
-import Footer from "./Footer";
-import Sidebar from "./components/Sidebar";
-import { AuthConsumer } from "./../../providers/AuthProvider";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    height: "100%",
-    minHeight: "100vh",
-    display: "flex",
-    flexDirection: "column",
-  },
-  shiftContent: {
-    paddingLeft: 240,
-  },
-  content: {
-    height: "100%",
-  },
-}));
-
-const Dashboard = (props) => {
-  const { children } = props;
-
-  const classes = useStyles();
-  const theme = useTheme();
-  const isDesktop = useMediaQuery(theme.breakpoints.up("lg"), {
-    defaultMatches: true,
-  });
-  const [openSidebar, setOpenSidebar] = useState(false);
-
-  const shouldOpenSidebar = isDesktop ? true : openSidebar;
-
-  const handleSidebarOpen = () => {
-    setOpenSidebar(true);
-  };
-  const handleSidebarClose = () => {
-    setOpenSidebar(false);
-  };
-  return (
-    <AuthConsumer>
-      {({ isAuth, logout, user }) => (
-        <div
-          className={clsx({
-            [classes.root]: true,
-            [classes.shiftContent]: isDesktop,
-          })}
-        >
-          <Header
-            onSidebarOpen={handleSidebarOpen}
-            logout={logout}
-            user={user}
-          />
-          <Sidebar
-            onClose={handleSidebarClose}
-            open={shouldOpenSidebar}
-            variant={isDesktop ? "persistent" : "temporary"}
-            isAuth={isAuth}
-            logout={logout}
-            user={user}
-          />
-
-          <Container maxWidth="lg" style={{ flex: 1 }}>
-            {children}
-          </Container>
-          <Footer />
-        </div>
-      )}
-    </AuthConsumer>
-  );
-};
-
-export default Dashboard;
+export { default } from "./Dashboard";
